[G38-23] Add authentication & accounts

+++ backend/account/src/profile/profile.controller.ts
@@ -1,9 +1,21 @@
-import { Controller, Get, Body, Param, Delete, Put } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
+import {
+  Controller,
+  Get,
+  Body,
+  Param,
+  Delete,
+  Put,
+  ForbiddenException,
+} from '@nestjs/common';
+import { ApiOkResponse, ApiTags } from '@nestjs/swagger';
 
 import { ProfileService } from './profile.service';
 import { UpdateProfileDto } from './dto/update-profile.dto';
-import { StatusResponseDto } from 'src/shared/status-response.dto';
+import { User } from '../users/user.entity';
+import { AuthUser } from '../shared/decorators/user.decorator';
+import { UseAuth } from '../shared/decorators/auth.decorator';
+import { Profile } from './profile.entity';
+import { JwtAuthGuard } from '../auth/guard/jwt-auth.guard';
 
 @ApiTags('Profile')
 @Controller('profile')
@@ -11,36 +23,54 @@ export class ProfileController {
   constructor(private readonly profileService: ProfileService) {}
 
   /**
+   * Get profile of authenticated user
+   */
+  @ApiOkResponse({ type: Profile })
+  @UseAuth(JwtAuthGuard)
+  @Get('me')
+  findMe(@AuthUser() requester: User): Promise<Profile> {
+    return this.profileService.findOne(requester.id);
+  }
+
+  /**
    * Get a user profile
    */
+  @ApiOkResponse({ type: Profile })
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Param('id') id: string): Promise<Profile> {
     return this.profileService.findOne(id);
   }
 
   /**
    * Update a user profile
    */
+  @UseAuth(JwtAuthGuard)
   @Put(':id')
-  update(@Param('id') id: string, @Body() updateProfileDto: UpdateProfileDto) {
-    // if (id !== requester.id) {
-    //   throw new UnauthorizedException('Not allowed to update profile');
-    // }
-    return this.profileService.update(id, updateProfileDto);
+  async update(
+    @AuthUser() requester: User,
+    @Param('id') id: string,
+    @Body() updateProfileDto: UpdateProfileDto
+  ): Promise<string> {
+    if (id !== requester.id) {
+      throw new ForbiddenException('Not allowed to update profile');
+    }
+    await this.profileService.update(id, updateProfileDto);
+    return 'Successfully updated profile';
   }
 
   /**
    * Delete a user profile
    */
+  @UseAuth(JwtAuthGuard)
   @Delete(':id')
-  async remove(@Param('id') id: string): Promise<StatusResponseDto> {
-    // if (id !== requester.id) {
-    //   throw new UnauthorizedException('Not allowed to delete profile');
-    // }
+  async remove(
+    @AuthUser() requester: User,
+    @Param('id') id: string
+  ): Promise<string> {
+    if (id !== requester.id) {
+      throw new ForbiddenException('Not allowed to delete profile');
+    }
     await this.profileService.remove(id);
-    return {
-      success: true,
-      message: 'Successfully deleted profile',
-    };
+    return 'Successfully deleted profile';
   }
 }
